Move the tax attribution result to the model
