+++ cgt_calc/model.py
@@ -8,7 +8,7 @@ from decimal import Decimal
 from enum import Enum
 import re
 import sys
-from typing import TYPE_CHECKING, Final, Self, override
+from typing import TYPE_CHECKING, Final, NamedTuple, Self, override
 
 from colorama import Style
 
@@ -212,6 +212,20 @@ ExcessReportedIncomeLog = dict[datetime.date, dict[str, ExcessReportedIncome]]
 ExcessReportedIncomeDistributionLog = dict[
     datetime.date, dict[str, ExcessReportedIncomeDistribution]
 ]
+
+
+class DividendTaxAttribution(NamedTuple):
+    """Withheld tax sorted by whether a dividend was found to carry it."""
+
+    matched: ForeignAmountLog
+    """Keyed by the symbol and date of the dividend the tax was taken from."""
+
+    unmatched: ForeignAmountLog
+    """Keyed by the symbol and date of the withholding itself.
+
+    No dividend row can carry this tax, but the broker took it all the same,
+    so the summary reports it against the year it was taken in.
+    """
 
 
 class ActionType(Enum):
