+++ tests/general/test_dividend_tax_attribution.py
@@ -255,7 +255,7 @@ def test_tax_two_payments_could_claim_is_left_out(
 
     assert len(warnings) == 1
     assert "the dividend on 2024-06-03 or 2024-07-03" in warnings[0]
-    assert "Date it on the one it belongs to" in warnings[0]
+    assert "Date it in the export on the one it belongs to" in warnings[0]
 
 
 def test_further_withholding_two_payments_could_claim_is_left_out(
